Add feature to enable docs for docs.rs
The feature "docs-only" removes the need to link and call LHAPDF, which
is needed to build the documentation of this crate without having the
dependency installed. This feature is enabled by default on docs.rs,
which fails for the exact reason.

This workaround was copied from cpp_build's Cargo.toml.